added named interface

+++ pkg/providers/activity/strava/route_test.go
@@ -21,7 +21,7 @@ func Test_Route(t *testing.T) {
 	rte, err := client.Route.Route(ctx, 26587226)
 	a.NoError(err)
 	a.NotNil(rte)
-	a.Equal(26587226, rte.ID)
+	a.Equal(int64(26587226), rte.ID)
 }
 
 func Test_Routes(t *testing.T) {
